Fuse subticker with clock
