+++ NewtonianKitTests/Time/ClockTests.swift
@@ -10,11 +10,31 @@ import Testing
 @testable import NewtonianKit
 
 struct ClockTests {
-  private lazy var subticker = VirtualSubticker()
-  private lazy var clock = Clock(subticker: subticker)
+  private var clock = Clock()
 
   init() async {
     await clock.start()
+  }
+
+  @Test func elapsedTimeIncreasesPerSubtick() async throws {
+    await clock.advanceTime(by: .subticks(2))
+    #expect(await clock.elapsedTime == .subticks(2))
+    await clock.stop()
+  }
+
+  @Test func doesNotAdvanceTimeWhenPaused() async throws {
+    await clock.pause()
+    await clock.advanceTime(by: .subticks(2))
+    #expect(await clock.elapsedTime == .zero)
+    await clock.stop()
+  }
+
+  @Test func performsPendingTimeAdvancementsWhenResumed() async throws {
+    await clock.pause()
+    await clock.advanceTime(by: .subticks(2))
+    await clock.start()
+    #expect(await clock.elapsedTime == .subticks(2))
+    await clock.stop()
   }
 
   @Test(arguments: [
@@ -23,7 +43,7 @@ struct ClockTests {
   ])
   mutating func adds(onTickListeners: [CountingOnTickListener]) async throws {
     for listener in onTickListeners { let _ = await clock.add(onTickListener: listener) }
-    await subticker.advance(by: .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
     for listener in onTickListeners { #expect(listener.count == 3) }
     await clock.stop()
   }
@@ -35,25 +55,25 @@ struct ClockTests {
   mutating func removes(onTickListeners: [CountingOnTickListener]) async throws {
     let ids = await onTickListeners.map { listener in await clock.add(onTickListener: listener) }
     for id in ids { await clock.removeOnTickListener(identifiedAs: id) }
-    await subticker.advance(by: .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
     for listener in onTickListeners { #expect(listener.count == 0) }
     await clock.stop()
   }
 
   @Test mutating func pausesSubtickerUponPause() async throws {
     await clock.pause()
-    await subticker.advance(by: .ticks(2))
-    #expect(await subticker.elapsedTime == .zero)
-    await subticker.resume()
-    #expect(await subticker.elapsedTime == .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
+    #expect(await clock.elapsedTime == .zero)
+    await clock.start()
+    #expect(await clock.elapsedTime == .ticks(2))
   }
 
   @Test mutating func stopsSubtickerUponStop() async throws {
-    await subticker.advance(by: .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
     await clock.stop()
-    await subticker.advance(by: .ticks(2))
-    await subticker.resume()
-    #expect(await subticker.elapsedTime == .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
+    await clock.start()
+    #expect(await clock.elapsedTime == .ticks(2))
   }
 
   @Test(arguments: [
@@ -63,7 +83,7 @@ struct ClockTests {
   mutating func removesUponStop(onTickListeners: [CountingOnTickListener]) async throws {
     for listener in onTickListeners { let _ = await clock.add(onTickListener: listener) }
     await clock.stop()
-    await subticker.advance(by: .ticks(2))
+    await clock.advanceTime(by: .ticks(2))
     for listener in onTickListeners { #expect(listener.count == 0) }
   }
 }
